--- fe/fe-core/src/main/java/org/apache/doris/common/proc/LoadErrorHubProcNode.java
@@ -1,47 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package org.apache.doris.common.proc;
-
-import org.apache.doris.catalog.Env;
-import org.apache.doris.load.LoadErrorHub;
-
-import com.google.common.collect.ImmutableList;
-
-public class LoadErrorHubProcNode implements ProcNodeInterface {
-    public static final ImmutableList<String> TITLE_NAMES = new ImmutableList.Builder<String>()
-            .add("Type").add("Properties")
-            .build();
-
-    private Env env;
-
-    public LoadErrorHubProcNode(Env env) {
-        this.env = env;
-    }
-
-    @Override
-    public ProcResult fetchResult() {
-        BaseProcResult result = new BaseProcResult();
-        result.setNames(TITLE_NAMES);
-        LoadErrorHub.Param param = env.getLoadInstance().getLoadErrorHubInfo();
-        if (param != null) {
-            result.addRow(param.getInfo());
-        }
-
-        return result;
-    }
-}
