+++ fe/fe-core/src/main/java/org/apache/doris/load/LoadErrorHub.java
@@ -17,10 +17,15 @@
 
 package org.apache.doris.load;
 
+import org.apache.doris.catalog.Env;
+import org.apache.doris.catalog.FsBroker;
 import org.apache.doris.common.io.Text;
 import org.apache.doris.common.io.Writable;
+import org.apache.doris.common.util.PrintableMap;
+import org.apache.doris.thrift.TBrokerErrorHubInfo;
 import org.apache.doris.thrift.TErrorHubType;
-import org.apache.doris.thrift.TLoadErrorHubInfo;
+import org.apache.doris.thrift.TMysqlErrorHubInfo;
+import org.apache.doris.thrift.TNetworkAddress;
 
 import com.google.common.base.MoreObjects;
 import com.google.common.base.MoreObjects.ToStringHelper;
@@ -37,7 +42,103 @@ import java.util.List;
 import java.util.Map;
 
 public abstract class LoadErrorHub {
-    private static final Logger LOG = LogManager.getLogger(LoadErrorHub.class);
+
+    public static class MysqlParam implements Writable {
+        private String host;
+        private int port;
+        private String user;
+        private String passwd;
+        private String db;
+        private String table;
+
+        public MysqlParam() {
+            host = "";
+            port = 0;
+            user = "";
+            passwd = "";
+            db = "";
+            table = "";
+        }
+
+        public MysqlParam(String host, int port, String user, String passwd, String db, String table) {
+            this.host = host;
+            this.port = port;
+            this.user = user;
+            this.passwd = passwd;
+            this.db = db;
+            this.table = table;
+        }
+
+        public String getBrief() {
+            Map<String, String> briefMap = Maps.newHashMap();
+            briefMap.put("host", host);
+            briefMap.put("port", String.valueOf(port));
+            briefMap.put("user", user);
+            briefMap.put("password", passwd);
+            briefMap.put("database", db);
+            briefMap.put("table", table);
+            PrintableMap<String, String> printableMap = new PrintableMap<>(briefMap, "=", true, false, true);
+            return printableMap.toString();
+        }
+
+        @Override
+        public void write(DataOutput out) throws IOException {
+            Text.writeString(out, host);
+            out.writeInt(port);
+            Text.writeString(out, user);
+            Text.writeString(out, passwd);
+            Text.writeString(out, db);
+            Text.writeString(out, table);
+        }
+
+        public void readFields(DataInput in) throws IOException {
+            host = Text.readString(in);
+            port = in.readInt();
+            user = Text.readString(in);
+            passwd = Text.readString(in);
+            db = Text.readString(in);
+            table = Text.readString(in);
+        }
+    }
+    
+
+    public static class BrokerParam implements Writable {
+        private String brokerName;
+        private String path;
+        private Map<String, String> prop = Maps.newHashMap();
+
+        // for persist
+        public BrokerParam() {
+        }
+
+        public BrokerParam(String brokerName, String path, Map<String, String> prop) {
+            this.brokerName = brokerName;
+            this.path = path;
+            this.prop = prop;
+        }
+
+        @Override
+        public void write(DataOutput out) throws IOException {
+            Text.writeString(out, brokerName);
+            Text.writeString(out, path);
+            out.writeInt(prop.size());
+            for (Map.Entry<String, String> entry : prop.entrySet()) {
+                Text.writeString(out, entry.getKey());
+                Text.writeString(out, entry.getValue());
+            }
+        }
+
+        public void readFields(DataInput in) throws IOException {
+            brokerName = Text.readString(in);
+            path = Text.readString(in);
+            int size = in.readInt();
+            for (int i = 0; i < size; i++) {
+                String key = Text.readString(in);
+                String val = Text.readString(in);
+                prop.put(key, val);
+            }
+        }
+    }
 
     public static final String MYSQL_PROTOCOL = "MYSQL";
     public static final String BROKER_PROTOCOL = "BROKER";
@@ -48,135 +149,14 @@ public abstract class LoadErrorHub {
         NULL_TYPE
     }
 
-    public class ErrorMsg {
-        private long jobId;
-        private String msg;
-
-        public ErrorMsg(long id, String message) {
-            jobId = id;
-            msg = message;
-        }
-
-        public long getJobId() {
-            return jobId;
-        }
-
-        public String getMsg() {
-            return msg;
-        }
-    }
-
     public static class Param implements Writable {
         private HubType type;
-        private MysqlLoadErrorHub.MysqlParam mysqlParam;
-        private BrokerLoadErrorHub.BrokerParam brokerParam;
+        private MysqlParam mysqlParam;
+        private BrokerParam brokerParam;
 
         // for replay
         public Param() {
             type = HubType.NULL_TYPE;
-        }
-
-        public static Param createMysqlParam(MysqlLoadErrorHub.MysqlParam mysqlParam) {
-            Param param = new Param();
-            param.type = HubType.MYSQL_TYPE;
-            param.mysqlParam = mysqlParam;
-            return param;
-        }
-
-        public static Param createBrokerParam(BrokerLoadErrorHub.BrokerParam brokerParam) {
-            Param param = new Param();
-            param.type = HubType.BROKER_TYPE;
-            param.brokerParam = brokerParam;
-            return param;
-        }
-
-        public static Param createNullParam() {
-            Param param = new Param();
-            param.type = HubType.NULL_TYPE;
-            return param;
-        }
-
-        public HubType getType() {
-            return type;
-        }
-
-        public MysqlLoadErrorHub.MysqlParam getMysqlParam() {
-            return mysqlParam;
-        }
-
-        public BrokerLoadErrorHub.BrokerParam getBrokerParam() {
-            return brokerParam;
-        }
-
-        public String toString() {
-            ToStringHelper helper = MoreObjects.toStringHelper(this);
-            helper.add("type", type.toString());
-            switch (type) {
-                case MYSQL_TYPE:
-                    helper.add("mysql_info", mysqlParam.toString());
-                    break;
-                case NULL_TYPE:
-                    helper.add("mysql_info", "null");
-                    break;
-                default:
-                    Preconditions.checkState(false, "unknown hub type");
-            }
-
-            return helper.toString();
-        }
-
-        public TLoadErrorHubInfo toThrift() {
-            TLoadErrorHubInfo info = new TLoadErrorHubInfo();
-            switch (type) {
-                case MYSQL_TYPE:
-                    info.setType(TErrorHubType.MYSQL);
-                    info.setMysqlInfo(mysqlParam.toThrift());
-                    break;
-                case BROKER_TYPE:
-                    info.setType(TErrorHubType.BROKER);
-                    info.setBrokerInfo(brokerParam.toThrift());
-                    break;
-                case NULL_TYPE:
-                    info.setType(TErrorHubType.NULL_TYPE);
-                    break;
-                default:
-                    Preconditions.checkState(false, "unknown hub type");
-            }
-            return info;
-        }
-
-        public Map<String, Object> toDppConfigInfo() {
-            Map<String, Object> dppHubInfo = Maps.newHashMap();
-            dppHubInfo.put("type", type.toString());
-            switch (type) {
-                case MYSQL_TYPE:
-                    dppHubInfo.put("info", mysqlParam);
-                    break;
-                case BROKER_TYPE:
-                    Preconditions.checkState(false, "hadoop load do not support broker error hub");
-                    break;
-                case NULL_TYPE:
-                    break;
-                default:
-                    Preconditions.checkState(false, "unknown hub type");
-            }
-            return dppHubInfo;
-        }
-
-        public List<String> getInfo() {
-            List<String> info = Lists.newArrayList();
-            info.add(type.name());
-            switch (type) {
-                case MYSQL_TYPE:
-                    info.add(mysqlParam.getBrief());
-                    break;
-                case BROKER_TYPE:
-                    info.add(brokerParam.getBrief());
-                    break;
-                default:
-                    info.add("");
-            }
-            return info;
         }
 
         @Override
@@ -200,11 +180,11 @@ public abstract class LoadErrorHub {
             type = HubType.valueOf(Text.readString(in));
             switch (type) {
                 case MYSQL_TYPE:
-                    mysqlParam = new MysqlLoadErrorHub.MysqlParam();
+                    mysqlParam = new MysqlParam();
                     mysqlParam.readFields(in);
                     break;
                 case BROKER_TYPE:
-                    brokerParam = new BrokerLoadErrorHub.BrokerParam();
+                    brokerParam = new BrokerParam();
                     brokerParam.readFields(in);
                     break;
                 case NULL_TYPE:
@@ -215,7 +195,6 @@ public abstract class LoadErrorHub {
         }
     }
 
-    public abstract List<ErrorMsg> fetchLoadError(long jobId);
 
     public abstract boolean prepare();
